add PrestigePad (WIP)

+++ contracts/land/constants/EstateLiquidatorConstant.sol
@@ -6,6 +6,6 @@ library EstateLiquidatorConstant {
     uint256 internal constant UNANIMOUS_QUORUM_RATE = 1 ether;
     uint256 internal constant MAJORITY_QUORUM_RATE = 0.75 ether;
 
-    uint256 internal constant ADMISSION_DURATION = 30 days;
-    uint256 internal constant VOTING_DURATION = 30 days;
+    uint40 internal constant ADMISSION_DURATION = 30 days;
+    uint40 internal constant VOTING_DURATION = 30 days;
 }
